refactor: remove index in view directory
