+++ src/index.js
@@ -1,6 +1,6 @@
 import { initComponents } from "./init/initComponents.js";
 import { initEventListener } from "./init/initEventListener.js";
-import { initViews } from "./init/intiViews.js";
+import { initViews } from "./init/initViews.js";
 
 initViews();
 initComponents();
