--- src/view/Main/Column/Card/index.js
@@ -1,2 +1,0 @@
-export { CardView } from "./CardView";
-export { AddCardFormView, EditCardFormView } from "./CardFormView";
